// src/models/document-command.ts
export type DocumentCommand =
	| 'backColor'
	| 'bold'
	| 'contentReadOnly'
	| 'copy'
	| 'createLink'
	| 'cut'
	| 'decreaseFontSize'
	| 'defaultParagraphSeparator'
	| 'delete'
	| 'enableAbsolutePositionEditor'
	| 'enableInlineTableEditing'
	| 'enableObjectResizing'
	| 'fontName'
	| 'fontSize'
	| 'foreColor'
	| 'formatBlock'
	| 'forwardDelete'
	| 'heading'
	| 'highlightColor'
	| 'increaseFontSize'
	| 'indent'
	| 'insertBrOnReturn'
	| 'insertHorizontalRule'
	| 'insertHTML'
	| 'insertImage'
	| 'insertOrderedList'
	| 'insertUnorderedList'
	| 'insertParagraph'
	| 'insertText'
	| 'italic'
	| 'justifyCenter'
	| 'justifyFull'
	| 'justifyLeft'
	| 'justifyRight'
	| 'outdent'
	| 'paste'
	| 'redo'
	| 'removeFormat'
	| 'selectAll'
	| 'strikeThrough'
	| 'subscript'
	| 'superscript'
	| 'underline'
	| 'undo'
	| 'unlink'
	| 'useCSS'
	| 'styleWithCSS'
	| 'AutoUrlDetect';